bot/requirements.txt
# Core dependencies
discord.py>=2.0.0
aiohttp>=3.8.0
twitchio>=2.6.0
python-dotenv>=0.19.0
requests>=2.28.0

# Deployment and Web
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
gunicorn>=20.1.0
python-json-logger>=2.0.0
aiohttp-cors>=0.7.0  # Added for health check server

# Monitoring
psutil>=5.9.0
